Rename to sole. Add ci and release workflows.
