+++ README.md
@@ -1,13 +1,13 @@
-# Knit 🧶
+# Sole 🥒
 
-[![Build Status](https://github.com/Solenya-AIaaS/knit/actions/workflows/ci.yml/badge.svg)](https://github.com/Solenya-AIaaS/knit/actions)
+[![Build Status](https://github.com/Solenya-AIaaS/sole/actions/workflows/ci.yml/badge.svg)](https://github.com/Solenya-AIaaS/sole/actions)
 [![Ruff](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/astral-sh/ruff/main/assets/badge/v2.json)](https://github.com/astral-sh/ruff)
-[![Python Version](https://img.shields.io/badge/python-3.12%2B-blue)](https://github.com/Solenya-AIaaS/knit)
+[![Python Version](https://img.shields.io/badge/python-3.12%2B-blue)](https://github.com/Solenya-AIaaS/sole)
 [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
 
-**Knit** is a tool for dynamically generating and maintaining markdown files using references to commands and external sources.
+**Sole** is a tool for dynamically generating and maintaining markdown files using references to commands and external sources.
 
-## 🤖 Why Knit?
+## 🤖 Why Sole?
 
 We are entering the **Agentic Era**. 
 
@@ -15,21 +15,21 @@ In this new paradigm, documentation is no longer just for humans, it's the prima
 
 The quality of an agent's output can degrade significantly if there is outdated or false context in your documentation. 
 
-**Knit** solves this by allowing you to "knit" your documentation together from the source of truth. Instead of copy-pasting output, you define *how* to generate it, ensuring your docs are always 100% accurate.
+**Sole** solves this by allowing you to generate your documentation from the source of truth. Instead of copy-pasting output, you define *how* to generate it, ensuring your docs are always 100% accurate.
 
 ## 🏗️ Why not Quarto?
-**DRY (Don't Repeat Yourself)**: Quarto is fantastic, but often relies on templates which can introduce duplication. Knit is designed to inject truth directly into your existing markdown files without the need for a complex build chain or intermediate template files.
+**DRY (Don't Repeat Yourself)**: Quarto is fantastic, but often relies on templates which can introduce duplication. Sole is designed to inject truth directly into your existing markdown files without the need for a complex build chain or intermediate template files.
 
-**Simplicity**: Quarto is a large, powerful ecosystem. Knit is a small, sharp tool designed specifically for maintaining context for Agents. It does one thing and does it well.
+**Simplicity**: Quarto is a large, powerful ecosystem. Sole is a small, sharp tool designed specifically for maintaining context for Agents. It does one thing and does it well.
 
 ## 🔒 Safety First
-**No Arbitrary Execution**: Unlike tools that allow running arbitrary bash commands in your documentation (an agentic security nightmare), Knit relies on typed, checked **Extensions**.
+**No Arbitrary Execution**: Unlike tools that allow running arbitrary bash commands in your documentation (an agentic security nightmare), Sole relies on typed, checked **Extensions**.
 
 **Secure by Design**: You define exactly what code can run via extensions. This prevents agents or malicious actors from using your documentation build system as a privilege escalation vector.
 
 ## ✨ Extensions
 
-Knit is designed around **Extensions**. It comes with powerful built-ins, but you can easily write your own in Python.
+Sole is designed around **Extensions**. It comes with powerful built-ins, but you can easily write your own in Python.
 
 ### 🌳 Tree Extension
 
@@ -42,7 +42,7 @@ Automatically visualize your directory structure. It's smart enough to pull desc
 .
 ├── dist
 ├── src
-│   └── knit
+│   └── sole
 └── tests
     └── __pycache__
 \```
@@ -72,7 +72,7 @@ Add your own extensions by defining a Python function.
 
 ```python
 # scripts/extensions.py
-from knit import register_extension
+from sole import register_extension
 
 @register_extension("HELLO")
 def hello(content, options, file_path):
@@ -82,13 +82,13 @@ def hello(content, options, file_path):
 
 Usage:
 ```markdown
-<!-- docs HELLO name="Knit" -->
+<!-- docs HELLO name="Sole" -->
 
-Hello Knit!
+Hello Sole!
 
 <!-- /docs -->
 ```
 
 ## 📈 Star History
 
-[![Star History Chart](https://api.star-history.com/svg?repos=Solenya-AIaaS/knit&type=Date)](https://star-history.com/#Solenya-AIaaS/knit&Date)
+[![Star History Chart](https://api.star-history.com/svg?repos=Solenya-AIaaS/sole&type=Date)](https://star-history.com/#Solenya-AIaaS/sole&Date)
