+++ src/sole/main.py
@@ -3,14 +3,14 @@ from pathlib import Path
 from typing import Annotated
 from rich.console import Console
 
-from knit import __version__
-from knit.core import process_content, clear_content
-from knit.registry import get_extension
+from sole import __version__
+from sole.core import process_content, clear_content
+from sole.registry import get_extension
 # Importing these modules registers the extensions via decorators
-import knit.extensions.tree
-import knit.extensions.just
+import sole.extensions.tree
+import sole.extensions.just
 
-app = typer.Typer(help="Knit: Auto-sync dynamic content in markdown files")
+app = typer.Typer(help="Sole: Auto-sync dynamic content in markdown files")
 console = Console()
 
 def load_extensions():
@@ -22,15 +22,15 @@ def load_extensions():
 @app.callback()
 def callback():
     """
-    Knit: A tool for maintaining dynamic markdown documentation.
+    Sole: A tool for maintaining dynamic markdown documentation.
     """
     load_extensions()
 
 
 @app.command()
 def version():
-    """Show the version of knit."""
-    console.print(f"Knit version: {__version__}")
+    """Show the version of sole."""
+    console.print(f"Sole version: {__version__}")
 
 
 @app.command()
