+++ tests/smoke_test.py
@@ -1,0 +1,5 @@
+import sole
+import sys
+
+print(f"Successfully imported sole version: {sole.__version__}")
+sys.exit(0)
